src/requirements.txt
einops
diffimg
pillow
python-dotenv
torch 
torchvision 
torchaudio
transformers
